printf

+++ _print_i_integer.c
@@ -16,7 +16,7 @@ if (num < 0)
 count += _putchar('-');
 num = -num;
 }
-if (num == 0 || num == NULL)
+if (num == 0 || num == (int) NULL)
 {
 count += _putchar('0');
 return (count);
